refactor(marketing): split the leaf playbooks into their own module (EP-5CC9C184)
Adding the leaf tier pushed marketing-playbooks.ts past the substrate hotspot
threshold (63 -> 64 production module hotspots). The baseline was not loosened.

Split rather than trimmed, because the two tables have different growth curves:
the category table is a fixed set of 24 industries, while the leaf table gains
an entry every time an archetype's category answer turns out to be wrong for
it. Keeping them in one file guarantees the ceiling is hit again on the next
archetype.

Parent module is back to 778 LOC; the resolver is re-exported so no caller
changes. 35 test files / 327 tests still pass.

Design-Grounding-Decision: no-contract-change (file split to satisfy the substrate hotspot ratchet; the resolver is re-exported so every caller and the resolution order are byte-for-byte unchanged)
